pyproject.toml
[project]
name = "compliance-research-assistant"
version = "0.1.0"
description = "Compliance Research Agent with RAG and web search"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=4.0.0",
    "tavily-python>=0.3.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.26.0",
]

